Break out D1/D2Categories into their own files
