+++ src/app/destiny1/d1-bucket-categories.ts
@@ -1,0 +1,19 @@
+export const D1Categories = {
+  Postmaster: ['LostItems', 'SpecialOrders', 'Messages'],
+  Weapons: ['Class', 'Primary', 'Special', 'Heavy'],
+  Armor: ['Helmet', 'Gauntlets', 'Chest', 'Leg', 'ClassItem'],
+  General: [
+    'Artifact',
+    'Ghost',
+    'Consumable',
+    'Material',
+    'Ornaments',
+    'Emblem',
+    'Shader',
+    'Emote',
+    'Ship',
+    'Vehicle',
+    'Horn',
+  ],
+  Progress: ['Bounties', 'Quests', 'Missions'],
+};
