+++ src/app/destiny1/d1-buckets.ts
@@ -2,26 +2,7 @@ import _ from 'lodash';
 import { getDefinitions } from './d1-definitions';
 import { InventoryBuckets, InventoryBucket } from '../inventory/inventory-buckets';
 import { BucketCategory } from 'bungie-api-ts/destiny2';
-
-export const D1Categories = {
-  Postmaster: ['LostItems', 'SpecialOrders', 'Messages'],
-  Weapons: ['Class', 'Primary', 'Special', 'Heavy'],
-  Armor: ['Helmet', 'Gauntlets', 'Chest', 'Leg', 'ClassItem'],
-  General: [
-    'Artifact',
-    'Ghost',
-    'Consumable',
-    'Material',
-    'Ornaments',
-    'Emblem',
-    'Shader',
-    'Emote',
-    'Ship',
-    'Vehicle',
-    'Horn',
-  ],
-  Progress: ['Bounties', 'Quests', 'Missions'],
-};
+import { D1Categories } from './d1-bucket-categories';
 
 // A mapping from the bucket hash to DIM item types
 const bucketToType = {
