+++ src/app/destiny2/d2-bucket-categories.ts
@@ -1,0 +1,7 @@
+export const D2Categories = {
+  Postmaster: ['Engrams', 'LostItems', 'Messages', 'SpecialOrders'],
+  Weapons: ['Class', 'Kinetic', 'Energy', 'Power', 'SeasonalArtifacts'],
+  Armor: ['Helmet', 'Gauntlets', 'Chest', 'Leg', 'ClassItem'],
+  General: ['Ghost', 'ClanBanners', 'Vehicle', 'Ships', 'Emblems', 'Finishers'],
+  Inventory: ['Consumables', 'Modifications', 'Shaders'],
+};
